feat: add GlobalSeo component and rename Seo to PageSeo
