+++ src/app.html
@@ -7,12 +7,7 @@
 		<link rel="icon" href="/favicon-96x96.png" type="image/png" sizes="96x96" />
 		<link rel="apple-touch-icon" href="/apple-touch-icon.png" />
 		<link rel="manifest" href="/site.webmanifest" />
-		<title>Take Shelter Now</title>
 		<meta name="viewport" content="width=device-width, initial-scale=1" />
-		<meta
-			name="description"
-			content="Find nearby emergency shelters fast. Serving the Springfield, MO region with real-time shelter locations and directions."
-		/>
 		<link rel="preconnect" href="https://fonts.googleapis.com" />
 		<link rel="preconnect" href="https://fonts.gstatic.com" crossorigin />
 		<link
