+++ src/lib/components/seo/PageSeo.svelte
