+++ src/lib/components/seo/index.ts
@@ -1,2 +1,3 @@
+export { default as GlobalSeo } from './GlobalSeo.svelte';
 export { default as JsonLd } from './JsonLd.svelte';
-export { default as Seo } from './Seo.svelte';
+export { default as PageSeo } from './PageSeo.svelte';
