+++ src/routes/+layout.svelte
@@ -3,6 +3,7 @@
 	import { onNavigate } from '$app/navigation';
 	import { page } from '$app/state';
 	import { config } from '$lib/config';
+	import { SITE_TITLE, siteUrl } from '$lib/seo';
 	import { AppShell } from '$lib/components/layout';
 	import { Toaster } from '$lib/components/ui/sonner';
 	import type { GeoPoint } from '$lib/geo';
@@ -94,26 +95,13 @@
 	}
 
 	const robotsContent = config.allowIndexing ? undefined : 'noindex, nofollow';
-	const siteTitle = 'Take Shelter Now';
-	const siteDescription = 'Find a safe place fast during weather emergencies.';
-	const defaultImageUrl = `${config.siteUrl}/og.png`;
+	let canonicalUrl = $derived(siteUrl(page.url.pathname));
 </script>
 
 <svelte:head>
-	<title>{siteTitle}</title>
-	<meta name="description" content={siteDescription} />
-	<meta property="og:site_name" content={siteTitle} />
-	<meta property="og:type" content="website" />
-	<meta property="og:title" content={siteTitle} />
-	<meta property="og:description" content={siteDescription} />
-	<meta property="og:image" content={defaultImageUrl} />
-	<meta property="og:image:width" content="1200" />
-	<meta property="og:image:height" content="630" />
-	<meta property="og:image:alt" content="Take Shelter Now social share image" />
-	<meta name="twitter:card" content="summary_large_image" />
-	<meta name="twitter:title" content={siteTitle} />
-	<meta name="twitter:description" content={siteDescription} />
-	<meta name="twitter:image" content={defaultImageUrl} />
+	<meta property="og:site_name" content={SITE_TITLE} />
+	<meta property="og:url" content={canonicalUrl} />
+	<link rel="canonical" href={canonicalUrl} />
 	{#if robotsContent}
 		<meta name="robots" content={robotsContent} />
 	{/if}
